Working back from global illumination model

+++ RayTracer/RayTracer/Vector3.cpp
@@ -10,7 +10,7 @@ Vector3::Vector3()
 	z = 0.0f;
 }
 
-Vector3::Vector3(float _x, float _y, float _z) : x(_x), y(_y), z(_z)
+Vector3::Vector3(double _x, double _y, double _z) : x(_x), y(_y), z(_z)
 {
 
 }
@@ -21,7 +21,7 @@ Vector3::~Vector3()
 
 }
 
-float Vector3::Distance(Vector3 posA, Vector3 posB)
+double Vector3::Distance(Vector3 posA, Vector3 posB)
 {
 	return 1.0f;
 }
